onFormSubmit function moved to NewThoughtForm component
